PCB layout 2 complete
layout 1 was incorrect as the connectors and LEDs were not located in the right place to correctly mate with the tube panel PCBA. When connectors and LEDs were moved, all the traces became scrambled. it was necessary to re-wire the entire board.
ESP32 pins were also remapped to improve ease of wiring